fix: null reference when compat registry key was never created

+++ XIVLauncher/Game/AdminCheck.cs
@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Collections.Generic;
 using System.Security.Principal;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using Microsoft.Win32;
-using Xceed.Wpf.Toolkit.PropertyGrid.Converters;
 
 namespace XIVLauncher.Game
 {
@@ -14,34 +9,37 @@ namespace XIVLauncher.Game
     {
         public static void RunCheck()
         {
-            var compatFlagKey = Registry.CurrentUser.OpenSubKey(
-                "Software\\Microsoft\\Windows NT\\CurrentVersion\\AppCompatFlags\\Layers", true);
-
-            var compatEntries = compatFlagKey.GetValueNames();
-
-            var entriesToFix = new Stack<string>();
-            foreach (var compatEntry in compatEntries)
-            {
-                if ((compatEntry.Contains("ffxiv") || compatEntry.Contains("XIVLauncher")) && ((string) compatFlagKey.GetValue(compatEntry, string.Empty)).Contains("RUNASADMIN"))
-                    entriesToFix.Push(compatEntry);
-            }
-
             var runningAsAdmin = new WindowsPrincipal(WindowsIdentity.GetCurrent())
                 .IsInRole(WindowsBuiltInRole.Administrator);
 
-            if (entriesToFix.Count > 0)
+            var compatFlagKey = Registry.CurrentUser.OpenSubKey(
+                "Software\\Microsoft\\Windows NT\\CurrentVersion\\AppCompatFlags\\Layers", true);
+
+            if (compatFlagKey != null)
             {
-                var result = MessageBox.Show("XIVLauncher and/or FINAL FANTASY XIV are set to run as administrator.\nThis can cause various issues, including addons failing to launch and hotkey applications failing to respond.\n\nDo you want to fix this issue automatically?", "XIVLauncher Problem", MessageBoxButton.OKCancel, MessageBoxImage.Exclamation);
+                var compatEntries = compatFlagKey.GetValueNames();
 
-                if (result != MessageBoxResult.OK) 
-                    return;
-
-                while (entriesToFix.Count > 0)
+                var entriesToFix = new Stack<string>();
+                foreach (var compatEntry in compatEntries)
                 {
-                    compatFlagKey.DeleteValue(entriesToFix.Pop());
+                    if ((compatEntry.Contains("ffxiv") || compatEntry.Contains("XIVLauncher")) && ((string) compatFlagKey.GetValue(compatEntry, string.Empty)).Contains("RUNASADMIN"))
+                        entriesToFix.Push(compatEntry);
                 }
 
-                return;
+                if (entriesToFix.Count > 0)
+                {
+                    var result = MessageBox.Show("XIVLauncher and/or FINAL FANTASY XIV are set to run as administrator.\nThis can cause various issues, including addons failing to launch and hotkey applications failing to respond.\n\nDo you want to fix this issue automatically?", "XIVLauncher Problem", MessageBoxButton.OKCancel, MessageBoxImage.Exclamation);
+
+                    if (result != MessageBoxResult.OK) 
+                        return;
+
+                    while (entriesToFix.Count > 0)
+                    {
+                        compatFlagKey.DeleteValue(entriesToFix.Pop());
+                    }
+
+                    return;
+                }
             }
 
             if (runningAsAdmin && !Properties.Settings.Default.HasComplainedAboutAdmin)
